Fixup warnings of -Wconversion

+++ be/src/column/object_column.cpp
@@ -228,7 +228,7 @@ void ObjectColumn<T>::fnv_hash(uint32_t* hash, uint32_t from, uint32_t to) const
     std::string s;
     for (uint32_t i = from; i < to; ++i) {
         s.resize(_pool[i].serialize_size());
-        int32_t size = _pool[i].serialize(reinterpret_cast<uint8_t*>(s.data()));
+        int32_t size = static_cast<int32_t>(_pool[i].serialize(reinterpret_cast<uint8_t*>(s.data())));
         hash[i] = HashUtil::fnv_hash(s.data(), size, hash[i]);
     }
 }
