+++ be/src/exec/pipeline/scan/olap_meta_scan_operator.h
@@ -55,8 +55,8 @@ public:
     ChunkSourcePtr create_chunk_source(MorselPtr morsel, int32_t chunk_source_index) override;
 
 private:
-    void attach_chunk_source(int32_t source_index) override {}
-    void detach_chunk_source(int32_t source_index) override {}
+    void attach_chunk_source(size_t source_index) override {}
+    void detach_chunk_source(size_t source_index) override {}
     bool has_shared_chunk_source() const override { return false; }
     ChunkPtr get_chunk_from_buffer() override;
     size_t num_buffered_chunks() const override;
