+++ be/src/exec/pipeline/set/union_passthrough_operator.cpp
@@ -32,11 +32,11 @@ Status UnionPassthroughOperator::push_chunk(RuntimeState* state, const ChunkPtr&
             // we should clone the src column instead of directly moving the src column.
             if (src_slot_item.ref_count > 1) {
                 auto dst_column = ColumnHelper::clone_column(dst_slot->type(), dst_slot->is_nullable(), src_column,
-                                                             src_chunk->num_rows());
+                                                             static_cast<uint32_t>(src_chunk->num_rows()));
                 _dst_chunk->append_column(std::move(dst_column), dst_slot->id());
             } else {
                 auto dst_column = ColumnHelper::move_column(dst_slot->type(), dst_slot->is_nullable(), src_column,
-                                                            src_chunk->num_rows());
+                                                            static_cast<uint32_t>(src_chunk->num_rows()));
                 _dst_chunk->append_column(std::move(dst_column), dst_slot->id());
             }
         }
@@ -49,7 +49,7 @@ Status UnionPassthroughOperator::push_chunk(RuntimeState* state, const ChunkPtr&
             auto* dst_slot = _dst_slots[i++];
             ColumnPtr& src_column = src_chunk->get_column_by_slot_id(src_slot->id());
             auto dst_column = ColumnHelper::move_column(dst_slot->type(), dst_slot->is_nullable(), src_column,
-                                                        src_chunk->num_rows());
+                                                        static_cast<uint32_t>(src_chunk->num_rows()));
             _dst_chunk->append_column(std::move(dst_column), dst_slot->id());
         }
     }
