+++ be/src/util/parse_util.cpp
@@ -88,7 +88,8 @@ int64_t ParseUtil::parse_mem_spec(const std::string& mem_spec_str, const int64_t
 
     if (multiplier != -1) {
         // Parse float - MB or GB
-        auto limit_val = StringParser::string_to_float<double>(mem_spec_str.data(), number_str_len, &result);
+        auto limit_val =
+                StringParser::string_to_float<double>(mem_spec_str.data(), static_cast<int>(number_str_len), &result);
 
         if (result != StringParser::PARSE_SUCCESS) {
             return -1;
@@ -97,7 +98,8 @@ int64_t ParseUtil::parse_mem_spec(const std::string& mem_spec_str, const int64_t
         bytes = multiplier * limit_val;
     } else {
         // Parse int - bytes or percent
-        auto limit_val = StringParser::string_to_int<int64_t>(mem_spec_str.data(), number_str_len, &result);
+        auto limit_val =
+                StringParser::string_to_int<int64_t>(mem_spec_str.data(), static_cast<int>(number_str_len), &result);
 
         if (result != StringParser::PARSE_SUCCESS) {
             return -1;
